Move BufferedConsole to com.meti.lib.util
